+++ src/states/auto-scaling-group.state.ts
@@ -1,6 +1,3 @@
 export type AutoScalingGroupState = {
   name: string;
-  desired: number;
-  min: number;
-  max: number;
 };